+++ src/de/spreclib/model/fixationtime/FixationTime.java
@@ -33,8 +33,7 @@ public enum FixationTime implements ISprecPart, IListProvideable {
     return this.codePart;
   }
 
-
-  public boolean hasValue(int durationMinutes) {
+  public boolean hasValue(long durationMinutes) {
 
     if (this.lowerBoundMinutes == null && this.upperBoundMinutes == null) {
       return false;
